requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
exceptiongroup==1.3.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
